Move headless from cmdtree internal to internal
