+++ cmd/state/internal/cmdtree/activate.go
@@ -3,8 +3,8 @@ package cmdtree
 import (
 	"github.com/spf13/viper"
 
-	"github.com/ActiveState/cli/cmd/state/internal/headless"
 	"github.com/ActiveState/cli/internal/captain"
+	"github.com/ActiveState/cli/internal/headless"
 	"github.com/ActiveState/cli/internal/locale"
 	"github.com/ActiveState/cli/internal/primer"
 	"github.com/ActiveState/cli/internal/prompt"
