+++ cmd/state/internal/cmdtree/packages.go
@@ -1,8 +1,8 @@
 package cmdtree
 
 import (
-	"github.com/ActiveState/cli/cmd/state/internal/headless"
 	"github.com/ActiveState/cli/internal/captain"
+	"github.com/ActiveState/cli/internal/headless"
 	"github.com/ActiveState/cli/internal/locale"
 	"github.com/ActiveState/cli/internal/primer"
 	"github.com/ActiveState/cli/internal/runners/packages"
